upgrade migration code written but not very functional yet;
the only time we would need to reinstall an upgrade is when the container that holds an upgrade had an AE and thus was duplicated (upgrades themselves cannot have AEs in our current system). This unfortunately does not work correctly. For some reason the AE-less upgrade is being duplicated. Also, when cyberware has an effect and has things installed into it, things also fail. committing so others can help me lol