Moved the stackable item code out of a hook
It doesn't make sense to do it there.

When an item is asked to be created on an actor, it should be decided if
Foundry should create the item or update the existing item.

Moved the code to createEmbeddedDocuments(). This allows for a consistent
return from createEmbeddedDocuments of affected document (either new or updated).