make isInstalled/installedIn getters;
This may be controversial; the obvious
disadvantage is that the way i am doing
it has a high-ish computational cost. But there are some advantages too.
When darin wrote the installed item
code, he said that installedItems.list was the "source of truth" for
what is installed where. However, the installedIn and isInstalled
properties were assigned independently of the aforementioned list.
Making these properties getters means that they are always accessible,
and will always correspond with
the source of truth, installedItems.list.