Suppress data validation during migration;
The migration system allows users to confirm/abort Migration before it
starts. This has the unintended consequence of allowing Foundry to load
and, more importantly, validate all of its documents. This leads to a
problem where documents which don't pass validation due to some change
in the schema, get put in "Foundry Jail" and are not included in their
typical collection (`game.actors`, using the Actors collection, as an
example). These documents need to be migrated to fix their outdated
schema, but they are not easily accessed from "Foundry Jail". The
simplest solution to this is to override the `validate` method for CPR
Documents and call the original Foundry function only if migration isn't
running, otherwise returning true.

This override can be found in `overrides.js` and is used in two places:
  1. *Actors and Items*: Actors and Items are already passed through a
  JavaScript Proxy object in `entity-factory.js` (which allows us to
  dynamically set the class for each Document [so that we aren't forced
  to have a single CPRItem Class type, for example]). If we add the
  override to the prototype of the `constructor` of the Proxy handler,
  we can override every Foundry Document that passes through this
  factory in one fell swoop.
  2. *System Data Models*: This is the property accessed at
  `someActor.system`. We have to override these data models' `validate`
  method as well because they don't just inherit the parent Document's
  `validate` method that we override in step 1.  This is achieved in
  `/datamodels/abstract.js` simply by overriding the method in the
  extending class.