Update `loadMixins` to  pass correct `this` value;
Basically, we were assigning the mixins to CPRItem prototype, which
means technically that the mixins were just being applied to everything.
By providing the item instance rather than the class prototype, the
functions now only actually get applied to the proper items. This was
masking an issue (#1116) on a function that never should have been
called in the first place (or at least been a more obvious error off the
bat)