Prevent `confirmContainerDelete` from showing for compendia Items.
This is because pack items don't have real installed item data, it's
simply epehemeral and stored in the flags. Thus there is nothing to
delete. This was causing issues where if the dialog is displayed, and
the option to delete installed items is selected, it would delete
corresponding items in-world.